rename wasmd to wasm client
