+++ wasmclient/client/log.go
